% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{slowa_domeny}
\alias{slowa_domeny}
\title{Rsemstrom}
\usage{
slowa_domeny(domena, rodzaj_slowa, kluczAPI)
}
\arguments{
\item{domena}{domena dla której pobieramy dane}

\item{rodzaj_slowa}{dostępne all (słowa na której obecnie widoczna jest domena), new (nowe słowa), lost (utracone słowa)}

\item{kluczAPI}{klucz API semstorm}
}
\description{
Pakiet umożliwia pobieranie danych z SemStorm (widoczność witryny w Top 50 i estymowany ruch)
}
\examples{
slowa_domeny('domena.pl',all,"9fxjh")
}
